requirements.txt
Flask
Gunicorn
faiss-cpu==1.7.4
pandas
numpy==1.26.4
nltk
emoji
contractions==0.1.73
Pillow
requests
torch==2.2.2
torchvision==0.17.2
transformers
scikit-learn==1.3.2
scipy
matplotlib
